First Commit - UI UX Class Portfolio
Showing the two main projects for this UI UX class and showing the other better things I have done in adobe etc. Use this as a lesson. We bought a design course because at our internship we were using adobe xd. We know the program, finish the design courses and make nice things.